requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
mistralai==0.2.0
jsonrpcserver==5.0.9
redis==5.0.1